# lib/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Api
  version: 1.0.0
  description: |
    Typed contract for the Complete iQ Carrier Audit API.

    Authenticated requests use the `sid` HTTP-only session cookie. The optional
    `X-Organization-Id` header selects one of the caller's organization
    memberships; otherwise the default (or earliest) membership is used.
    Organization-scoped lookups deliberately return `404` when a resource
    belongs to another tenant so that cross-tenant identifiers are not
    disclosed.

    Job-enqueueing operations that accept `X-Idempotency-Key` scope that value
    to the selected organization and operation inputs. Repeating an equivalent
    request returns the existing durable processing job and sets `duplicate`
    to `true` when that field is present.
servers:
  - url: /api
    description: Application API
security:
  - cookieAuth: []
...
paths:
...
  /carriers:
    get:
      operationId: listActiveCarriers
      tags: [Carriers]
      summary: List active carrier choices
      description: Authenticated carrier choices used by intake and reprocessing interfaces.
      responses:
        "200":
          description: Active carriers.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CarrierOption"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /carriers/all:
    get:
      operationId: listCarriers
      tags: [Carriers, Administration]
      summary: List all carrier rulesets
      description: Returns active and inactive global carrier rulesets. Requires global `admin`.
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
      responses:
        "200":
          description: All carrier ruleset records.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CarrierProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /carriers/{key}:
    parameters:
      - $ref: "#/components/parameters/CarrierKey"
      - $ref: "#/components/parameters/OrganizationId"
    get:
      operationId: getCarrier
      tags: [Carriers]
      summary: Get a carrier ruleset
      description: Returns one global carrier ruleset. Requires global `admin`.
      responses:
        "200":
          description: Carrier ruleset.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarrierProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      operationId: upsertCarrier
      tags: [Carriers, Administration]
      summary: Create a governed carrier ruleset version
      description: |
        Requires global `admin`. A false `active` value creates an immutable draft
        version; true creates and explicitly publishes a validated version while
        preserving the prior publication.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertCarrierRequest"
      responses:
        "200":
          description: Carrier ruleset stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarrierVersionMutationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      operationId: deleteCarrier
      tags: [Carriers, Administration]
      summary: Deactivate a carrier ruleset
      description: Requires global `admin`; publication history is preserved.
      responses:
        "200":
          description: Carrier ruleset deactivated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarrierDeactivationResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /carriers/{key}/versions:
    get:
      operationId: listCarrierVersions
      tags: [Carriers, Administration]
      summary: List immutable carrier ruleset versions
      description: Requires global `admin`.
      parameters:
        - $ref: "#/components/parameters/CarrierKey"
        - $ref: "#/components/parameters/OrganizationId"
      responses:
        "200":
          description: Version history and current-organization impact count.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarrierVersionHistory"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /carriers/{key}/versions/{versionId}/publish:
    post:
      operationId: publishCarrierVersion
      tags: [Carriers, Administration]
      summary: Approve and publish a draft ruleset version
      description: Requires global `admin`; validates the persisted draft before publication.
      parameters:
        - $ref: "#/components/parameters/CarrierKey"
        - $ref: "#/components/parameters/OrganizationId"
        - $ref: "#/components/parameters/CarrierVersionId"
      responses:
        "200":
          description: Published version.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarrierVersionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /carriers/{key}/versions/{versionId}/rollback:
    post:
      operationId: rollbackCarrierVersion
      tags: [Carriers, Administration]
      summary: Restore a historical publication as a new version
      description: Requires global `admin`; prior versions remain immutable.
      parameters:
        - $ref: "#/components/parameters/CarrierKey"
        - $ref: "#/components/parameters/OrganizationId"
        - $ref: "#/components/parameters/CarrierVersionId"
      responses:
        "200":
          description: Newly created rollback publication.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarrierVersionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /carriers/{key}/test:
    post:
      operationId: testCarrierVersion
      tags: [Carriers, Administration]
      summary: Run deterministic representative-claim ruleset preflight
      description: |
        Requires global `admin` and `claims:read`. This does not call the AI
        provider and does not mutate the representative claim.
      parameters:
        - $ref: "#/components/parameters/CarrierKey"
        - $ref: "#/components/parameters/OrganizationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarrierPreflightRequest"
      responses:
        "200":
          description: Ruleset compatibility and coverage result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarrierPreflightResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sid
      description: HTTP-only server session cookie, marked Secure outside development.
    inboundToken:
      type: apiKey
      in: query
      name: token
      description: Shared secret configured for the SendGrid Inbound Parse callback.
  parameters:
    OrganizationId:
      name: X-Organization-Id
      in: header
      required: false
      description: |
        Organization membership to use for this request. Omit to use the
        caller's default membership. An inaccessible value returns `403`.
      schema:
        type: string
        format: uuid
...
    CarrierKey:
      name: key
      in: path
      required: true
      description: Stable normalized carrier key.
      schema:
        type: string
        minLength: 1
...
  responses:
...
    Unauthorized:
      description: A valid session or organization membership is required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The caller lacks the selected organization or required permission.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: |
        Resource not found in the selected organization. Organization-scoped
        routes use this response for cross-tenant identifiers.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    TooManyRequests:
      description: Request rate limit exceeded.
      headers:
        RateLimit-Limit:
          schema:
            type: integer
        RateLimit-Remaining:
          schema:
            type: integer
        RateLimit-Reset:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalServerError:
      description: The server could not complete the request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    CarrierOption:
      type: object
      required: [key, displayName, logoUrl]
      properties:
        key:
          type: string
        displayName:
          type: string
        logoUrl:
          type: string
          nullable: true
      additionalProperties: false
...
    CarrierProfile:
      type: object
      required: [id, carrierKey, displayName, logoUrl, active, ruleset, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        carrierKey:
          type: string
        displayName:
          type: string
        logoUrl:
          type: string
          nullable: true
        active:
          type: boolean
        ruleset:
          $ref: "#/components/schemas/CarrierRuleset"
        sourceReferences:
          type: array
          items:
            $ref: "#/components/schemas/CarrierSourceReference"
        changeSummary:
          type: string
          nullable: true
        hasDraft:
          type: boolean
        latestVersion:
          allOf:
            - $ref: "#/components/schemas/CarrierRulesetVersion"
          nullable: true
        publishedVersion:
          allOf:
            - $ref: "#/components/schemas/CarrierRulesetVersion"
          nullable: true
        createdAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
      additionalProperties: false
    UpsertCarrierRequest:
      type: object
      required: [displayName, logoUrl, active, ruleset, sourceReferences, changeSummary]
      properties:
        displayName:
          type: string
          minLength: 1
        logoUrl:
          type: string
          nullable: true
        active:
          type: boolean
        ruleset:
          $ref: "#/components/schemas/CarrierRuleset"
        sourceReferences:
          type: array
          items:
            $ref: "#/components/schemas/CarrierSourceReference"
        changeSummary:
          type: string
          minLength: 1
      additionalProperties: false
...
    CarrierVersionMutationResponse:
      type: object
      required: [success, version]
      properties:
        success:
          type: boolean
        version:
          $ref: "#/components/schemas/CarrierRulesetVersion"
      additionalProperties: false
...